pyproject.toml
[project]
name = "2025-1-t01-lfbagyourdreams"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asgiref==3.8.1",
    "botocore>=1.39.4",
    "certifi==2025.6.15",
    "charset-normalizer==3.4.2",
    "coverage>=7.9.2",
    "deprecation==2.1.0",
    "django==5.2.3",
    "django-cors-headers>=4.7.0",
    "django-environ==0.12.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "environ>=1.0",
    "faker>=37.4.0",
    "gotrue==2.12.2",
    "h11==0.16.0",
    "h2==4.2.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "hyperframe==6.1.0",
    "idna==3.10",
    "iniconfig==2.1.0",
    "packaging==25.0",
    "pluggy==1.6.0",
    "postgrest==1.1.1",
    "psycopg2-binary==2.9.10",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pytest==8.4.1",
    "pytest-django>=4.11.1",
    "pytest-mock==3.14.1",
    "python-dateutil==2.9.0.post0",
    "realtime==2.5.3",
    "requests==2.32.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlparse==0.5.3",
    "storage3==0.12.0",
    "strenum==0.4.15",
    "supabase==2.16.0",
    "supafunc==0.10.1",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "websockets==15.0.1",
]
